requirements.txt
asgiref==3.8.1
Django==5.1.2
django-quill-editor==0.1.42
sqlparse==0.5.1
tzdata==2024.2
